requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

psycopg2-binary>=2.9.9
django-environ>=0.11.2

celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

gunicorn>=21.2.0

pytest-django>=4.7.0